refactor(loom): extract provider handoff orchestration
